backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
pydantic==2.5.0
python-jose[cryptography]==3.3.0
python-multipart==0.0.6
bcrypt==4.1.2
pytest==7.4.0
pytest-mock==3.12.0
PyJWT==2.8.0
httpx==0.27.0
sqlalchemy-utils==0.41.1